Revert "Nbgs tsg archlayerchange"
