+++ cmd/main_release.go
@@ -5,13 +5,14 @@ package main
 import (
 	"fmt"
 	cmn "ng_autocode_back_service/common/model"
+	startup "ng_autocode_back_service/internal/startup"
 	"os"
 )
 
 func main() {
 	fmt.Println("Hello World - Fixed deposit")
 	// Initialize the application
-	if err := Startup(cmn.BUILD_RELEASE, cmn.APP_FIXED_DEPOSIT); err != nil {
+	if err := startup.Startup(cmn.BUILD_RELEASE, cmn.APP_FIXED_DEPOSIT); err != nil {
 		fmt.Fprintf(os.Stderr, "Startup error: %v\n", err)
 		os.Exit(1)
 	}
