+++ internal/core/fd/list/service/listServices.go
@@ -1,0 +1,16 @@
+package service
+
+import (
+	"context"
+	dto "ng_autocode_back_service/internal/core/fd/list/service/model/dto"
+)
+
+func (g *serviceItem) GetFDAgentList(c context.Context, dtoRequest *dto.FDAgentListRequest) (dtoResponse []*dto.FDAgentListResponse, e error) {
+	// Call the data layer to get the FDAgent list
+	dtoResponse, err := g.dataL.GetFDAgentList(c, dtoRequest)
+	if err != nil {
+		return nil, err
+	}
+
+	return dtoResponse, nil
+}
